todo-backend: add support for configuring the default parser
